+++ MERN/webpack/prod.config.js
@@ -1,18 +1,20 @@
 var path = require('path');
 var webpack = require('webpack');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-var jsDestPath = '../dist';
+var jsDestPath = 'dist';
+const RESOLVED_DIST_PATH = path.join(__dirname, '../', jsDestPath);
 
 var webpackOptions = {
+  mode: 'production',
   watch: false,
   entry: [
     './client/client.js'
   ],
   output: {
-    path: __dirname + '/'+jsDestPath,
+    path: RESOLVED_DIST_PATH,
     filename: 'client.js',
-    chunkFilename: '[id].[chunkhash].js',
-    publicPath :  'http://localhost:3040/'
+    chunkFilename: './[id].[chunkhash].js'
   },
   module: {
     rules: [
@@ -28,25 +30,50 @@ var webpackOptions = {
       },
       { test: /\.css$/, use: ['style-loader', 'css-loader'] },
       { test: /\.scss$/, use: ['style-loader', 'css-loader', 'sass-loader']},
-      { test: /\.png/, use: ['url-loader'] },
-      { test: /\.jpg/, use: ['url-loader'] },
+      {
+        test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
+        use: [
+          {
+            loader: 'file-loader',
+            options: {
+              name: '[name].[ext]',
+              outputPath: 'fonts/'
+            }
+          }
+        ]
+      }, {
+        test: /\.(png|jpe?g|gif)$/,
+        use: [
+          {
+            loader: 'url-loader',
+            options: {
+              name: 'images/[name].[ext]'
+            }
+          }
+        ]
+      }
     ]
   },
   resolve: {
+    /**
+     * ### webpack: directly import the sub-folders like a module
+     */
     modules: ['node_modules', 'client'],
-    extensions: ['', '.js', '.json','.jsx','.css']
+    extensions: ['.js', '.json','.jsx','.scss','.css']
   },
-  devServer: {
-    contentBase: __dirname + '/client/',
-		noInfo: true, //  --no-info option
-		inline: true
-	},
   node: {
     fs: 'empty',
     net: 'empty',
     tls: 'empty'
   },
   plugins: [
+    new HtmlWebpackPlugin({
+      title: 'React app',
+      filename: 'index.html', // default value
+      template: 'index.html',
+      inject: true,
+      excludeChunks: [ 'server' ]
+    }),
     // new webpack.optimize.UglifyJsPlugin([{
     //   compress: {
     //     warning: false