+++ MERN/CHANGELOG.md
@@ -6,3 +6,11 @@ Added login page
 
 ## 2.0.0
 Used react 16, webpack 4 and latest version of other packages
+
+## 2.0.1
+Added eslint rules
+
+## 2.0.2
+1. Clean up prod and hot webpack config, server files, .env file.
+2. Fixed glitches in production version of client and server build.
+3. Pure static server